requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.9
click==8.2.1
filetype==1.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Js2Py_3.13==0.74.1
lxml==6.0.0
m3u8==6.0.0
MarkupSafe==3.0.2
pyjsparser==2.7.1
pyotp==2.9.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
SQLAlchemy==2.0.41
twikit==2.3.3
typing_extensions==4.14.0
tzlocal==5.3.1
webvtt-py==0.5.1
Werkzeug==3.1.3

twikit
python-dotenv

